Implementacao da Palicy para Task

+++ app/Http/Controllers/TasksController.php
@@ -8,26 +8,36 @@ use App\Http\Requests\TasksRequest as Request;
 class TasksController extends Controller
 {
 
-    public function index() {
+    public function index()
+    {
         return Task::all();
     }
 
-    public function store(Request $request) {
-        return Task::create($request->all());
+    public function store(Request $request)
+    {
+        $data = $request->all();
+        $data['user_id'] = \Auth::user()->id;
+        return Task::create($data);
     }
 
-    public function update(Request $request, Task $task) {
+    public function update(Request $request, Task $task)
+    {
+        $this->authorize('update', $task);
         $task->update($request->all());
         return $task;
     }
 
-    public function show(Task $task) {
+    public function show(Task $task)
+    {
+        $this->authorize('view', $task);
         return $task;
     }
 
-    public function destroy(Task $task) {
-         $task->delete();
-         return $task;
+    public function destroy(Task $task)
+    {
+        $this->authorize('delete', $task);
+        $task->delete();
+        return $task;
     }
